challenges/02-security-audit/requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
psycopg2-binary==2.9.6
bcrypt==4.0.1
PyJWT==2.7.0
python-dotenv==1.0.0
requests==2.31.0
